refactor `RegionMeta` to `RegionContext` to support multiple wirting to a specific table.
